requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.10.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
marshmallow==3.26.1
motor==3.7.0
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
openai==1.66.3
packaging==24.2
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
textblob==0.19.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
